requirements.txt
google-api-python-client>=2.88.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
mcp[cli]
pydantic>=2.0.0 